add makeCollection
The WebDAV spec requires that MKCOL fail if the intermediate (parent)
collection does not exist, and so it seems best to make this API reflect
that, while also making it easy to detect that problem.

hdav makes missing intermediate collections.

It'd be nice to have a makeCollectionRecursive that makes intermediate
collections and reuses a single http connection while doing so. But I was
not sure how to do that, since DAVState contains a single url, and this
would need to operate on multiple urls.