add support for large files
The current getContentM and putContentM interfaces buffer the whole file
content in memory.

However, http-client provides interfaces that can be used to avoid this
problem.

For putContentM, the problem is that, while a lazy bytestring is
provided to RequestBodyLBS, http-client needs to know the length of its
request body. So it buffers the whole thing to calculate its length.
(http-client's documentation could be more clear about this..)

Fixing putContentM is as simple as exposing an interface that lets the
caller provide a RequestBody. Then they can provide eg, a RequestBodyStream
streaming in the content of a file in constant memory.

For getContentM, things are tricker, because davRequest calls httpLbs,
which buffers the whole response body. So, the Request generation needs
to be factored out, and then davRequest can be left as-is, while
providing a withContentM that uses Network.HTTP.Client.withResponse.
The caller can then stream the response (eg out to a file) in constant
memory.