add inDAVLocation and getDAVLocation
inDAVLocation is useful for composing DAV monad actions that operate on
different locations within the same DAV repository.

getDAVLocation can be useful when writing code that might need to eg,
create the parent collection that will hold the current location.

Needed to pull in utf-8 string in order to convert from Request paths
to strings.